# packages/backend/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Card": {
				"properties": {
					"id": {
						"type": "string"
					},
					"columnId": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"columnId",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Column": {
				"properties": {
					"id": {
						"type": "string"
					},
					"columnName": {
						"type": "string"
					},
					"cards": {
						"items": {
							"$ref": "#/components/schemas/Card"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"columnName",
					"cards"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Column.Exclude_keyofColumn.id-or-cards__": {
				"properties": {
					"columnName": {
						"type": "string"
					}
				},
				"required": [
					"columnName"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Column.id-or-cards_": {
				"$ref": "#/components/schemas/Pick_Column.Exclude_keyofColumn.id-or-cards__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"ColumnCreationRequestDTO": {
				"$ref": "#/components/schemas/Omit_Column.id-or-cards_"
			},
			"Pick_Card.Exclude_keyofCard.id__": {
				"properties": {
					"columnId": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"columnId",
					"description"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Card.id_": {
				"$ref": "#/components/schemas/Pick_Card.Exclude_keyofCard.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CardCreationRequestDTO": {
				"$ref": "#/components/schemas/Omit_Card.id_"
			},
			"Partial_Card_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"columnId": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@app/backend",
		"version": "0.0.0",
		"contact": {}
	},
	"paths": {
		"/api/column": {
			"get": {
				"operationId": "GetColumns",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Column"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of all columns",
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateColumn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Column"
								}
							}
						}
					}
				},
				"description": "Create new column",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ColumnCreationRequestDTO"
							}
						}
					}
				}
			}
		},
		"/api/column/{columnId}": {
			"get": {
				"operationId": "GetColumnById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Column"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Retrieves column by columnID",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "columnId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteColumnById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Column"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "columnId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/card": {
			"get": {
				"operationId": "GetCards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Card"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of all cards",
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Card"
								}
							}
						}
					}
				},
				"description": "Create new card",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CardCreationRequestDTO"
							}
						}
					}
				}
			}
		},
		"/api/card/{cardId}": {
			"get": {
				"operationId": "GetCardById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Card"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Retrieves card by cardID",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cardId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "EditCardById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Card"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cardId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Card_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCardById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Card"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "cardId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}